+++ micro/src/org/cougaar/microedition/asset/SensorResource.java
@@ -21,7 +21,7 @@ public abstract class SensorResource extends Resource {
 
   public abstract void setChan(int c);
   public abstract void setUnits(String u);
-  public abstract double getValue();
+  public abstract long getValue();
 
   public SensorResource() {}
